+++ public/js/vue/store/mutations.js
@@ -29,7 +29,7 @@ export default {
     state.search.messageSearchError = false;
     state.search.messageSearchLoading = true;
   },
-  [types.RECEIVE_MESSAGE_SEARCH_SUCESS](state, searchResults) {
+  [types.RECEIVE_MESSAGE_SEARCH_SUCCESS](state, searchResults) {
     state.search.messageSearchError = false;
     state.search.messageSearchLoading = false;
     state.search.messageSearchResults = searchResults;
@@ -42,7 +42,12 @@ export default {
 
   [types.CHANGE_DISPLAYED_ROOM](state, newRoomId) {
     state.displayedRoomId = newRoomId;
+    state.hightLightedMessageId = null;
   },
+  [types.CHANGE_HIGHLIGHTED_MESSAGE_ID](state, newMessageId) {
+    state.hightLightedMessageId = newMessageId;
+  },
+
   [types.UPDATE_ROOM](state, newRoomState) {
     if (newRoomState.id) {
       const resultantRoomState = Object.assign(