+++ scripts/utils/update-room-lcowner.js
@@ -92,7 +92,7 @@ function logProgress() {
   console.log('[PROGRESS]', '\tprocessed:', PROCESSED, '\tupdated:', UPDATED);
 }
 
-// reponsible for running the procedure
+// responsible for running the procedure
 function run(rooms) {
   // increment stuff
   CALLED_RUN += 1;