+++ test/public-js/vue/store/actions-test.js
@@ -2,14 +2,21 @@
 
 const createState = require('../../../../public/js/vue/store/state').default;
 const types = require('../../../../public/js/vue/store/mutation-types');
+
+jest.mock('gitter-web-client-context');
+
 const actions = require('../../../../public/js/vue/store/actions');
 const testAction = require('./vuex-action-helper');
+const context = require('gitter-web-client-context');
 const appEvents = require('../../../../public/js/utils/appevents');
+
+const { createSerializedRoomFixture } = require('../fixture-helpers');
 
 describe('actions', () => {
   let state;
   beforeEach(() => {
     state = createState();
+    context.troupe.mockReset();
   });
 
   it('setInitialData', done => {
@@ -103,7 +110,7 @@ describe('actions', () => {
         state,
         [
           { type: types.REQUEST_MESSAGE_SEARCH },
-          { type: types.RECEIVE_MESSAGE_SEARCH_SUCESS, payload: null }
+          { type: types.RECEIVE_MESSAGE_SEARCH_SUCCESS, payload: null }
         ],
         [],
         done
@@ -132,7 +139,7 @@ describe('actions', () => {
 
   describe('changeDisplayedRoom', () => {
     it('room we do not know about will not send appEvents', async () => {
-      const payload = 900009;
+      const payload = '5cf8efbc4dfb4240048b768e';
 
       let appEventTriggered = false;
       appEvents.once('*', () => {
@@ -150,12 +157,12 @@ describe('actions', () => {
       expect(appEventTriggered).toEqual(false);
     });
 
-    it('fires appEvents to change rooms in the legacy part of the app', async () => {
-      const roomObject = {
-        id: 123456,
-        uri: 'foo/bar',
-        unreads: 4
-      };
+    it('when router-chat loaded, fires appEvents to change rooms in the legacy part of the app', async () => {
+      const roomObject = createSerializedRoomFixture('community/room1');
+
+      context.troupe.mockImplementation(function() {
+        return roomObject;
+      });
 
       state.roomMap[roomObject.id] = roomObject;
 
@@ -181,10 +188,50 @@ describe('actions', () => {
       await navigationEventFiredPromise;
       await vueChangeRoomEventFiredPromise;
     });
+
+    it('when on non-router-chat page, just redirects the page', async () => {
+      const roomObject = createSerializedRoomFixture('community/room1');
+
+      state.roomMap[roomObject.id] = roomObject;
+
+      window.location.assign = jest.fn();
+
+      await testAction(
+        actions.changeDisplayedRoom,
+        roomObject.id,
+        state,
+        [{ type: types.CHANGE_DISPLAYED_ROOM, payload: roomObject.id }],
+        []
+      );
+
+      expect(window.location.assign).toHaveBeenCalledWith(roomObject.url);
+    });
+  });
+
+  describe('jumpToMessageId', () => {
+    it('updates highlighted message ID and sends off appevent for legacy chat view backbone to consume', async () => {
+      const payload = '5cf8efbc4dfb4240048b768e';
+
+      const vuehightLightedMessageIdEventFiredPromise = new Promise(resolve => {
+        appEvents.on('vue:hightLightedMessageId', () => {
+          resolve();
+        });
+      });
+
+      await testAction(
+        actions.jumpToMessageId,
+        payload,
+        state,
+        [{ type: types.CHANGE_HIGHLIGHTED_MESSAGE_ID, payload: payload }],
+        []
+      );
+
+      await vuehightLightedMessageIdEventFiredPromise;
+    });
   });
 
   it('updateRoom', done => {
-    const payload = { id: 123456, unreads: 5 };
+    const payload = { id: '5cf8efbc4dfb4240048b768e', unreads: 5 };
     testAction(
       actions.updateRoom,
       payload,