+++ test/public-js/vue/store/getters-test.js
@@ -14,6 +14,27 @@ describe('getters', () => {
     state = createState();
   });
 
+  describe('displayedRoom', () => {
+    it('non-existent room is undefined', () => {
+      state.displayedRoomId = '000';
+
+      const displayedRoom = getters.displayedRoom(state);
+      expect(displayedRoom).toEqual(undefined);
+    });
+
+    it('returns full room object for displayedRoomId', () => {
+      const room1 = createSerializedRoomFixture('community/room1');
+      state.roomMap = {
+        [room1.id]: room1
+      };
+
+      state.displayedRoomId = room1.id;
+
+      const displayedRoom = getters.displayedRoom(state);
+      expect(displayedRoom).toEqual(room1);
+    });
+  });
+
   describe('displayedRooms', () => {
     it('when the user has not joined any rooms yet, then no rooms to display', () => {
       state.leftMenuState = 'all';
@@ -25,9 +46,9 @@ describe('getters', () => {
     it('when the left-menu is in "All conversations" state, shows all rooms', () => {
       state.leftMenuState = 'all';
 
-      const room1 = createSerializedRoomFixture('a/1');
-      const room2 = createSerializedRoomFixture('b/2');
-      const oneToOneRoom1 = createSerializedOneToOneRoomFixture('onetoone/1');
+      const room1 = createSerializedRoomFixture('community/room1');
+      const room2 = createSerializedRoomFixture('community/room2');
+      const oneToOneRoom1 = createSerializedOneToOneRoomFixture('onetoone/room1');
       state.roomMap = {
         [room1.id]: room1,
         [room2.id]: room2,
@@ -41,9 +62,9 @@ describe('getters', () => {
     it('when the left-menu is in "People"/one to one rooms state, only shows one to one rooms', () => {
       state.leftMenuState = 'people';
 
-      const room1 = createSerializedRoomFixture('a/1');
-      const room2 = createSerializedRoomFixture('b/2');
-      const oneToOneRoom1 = createSerializedOneToOneRoomFixture('onetoone/1');
+      const room1 = createSerializedRoomFixture('community/room1');
+      const room2 = createSerializedRoomFixture('community/room2');
+      const oneToOneRoom1 = createSerializedOneToOneRoomFixture('onetoone/room1');
       state.roomMap = {
         [room1.id]: room1,
         [room2.id]: room2,